Duck | Integrated with SQLCipher.
Had to make the Drishti DB name configurable, since the repository tests failed
without it. The onCreate() on the SQLCipher's SQLiteOpenHelper class is called
only once, when the DB does not exist. The tests failed because different tests
try to create different part of the DB at different times (for example,
SettingsRepositoryTest just initializes the "settings" table, and not the ec and
alert tables). So, after SettingsRepositoryTest finished, if the
AlertRepositoryTest started, the "alerts" table would not be present.

Created maven dependencies for the native libraries and SQLCipher. However, due
to the way IntelliJ creates the APK, the native libraries are duplicated in
src/main/native as well.