Harshit & Duck | Fixed long-standing ignore test.
There's a bug in the android emulator for 2.3.3. So, we cannot write a
test against any of our webview activities. This is why the ignored test
was failing, and the login activity tests were failing (only on the
emulator). Most login activity tests are now fixed by removing the part
which goes to HomeActivity.